Arbiter: optionally stop system if Actor/Arbiter panics
When an Actor panics in an Arbiter, the Arbiter thread stops but no
other action is taken, the rest of the program continues to run. Only if
a sender is using try_send or send will they notice the failure. This
can lead to surprising circumstances where a program will appear to be
running but instead silently misbehaving or ready to explode on the next
message to that Actor (as a panic while processing the sent message does
_not_ fail initially).

Given that a panic is usually indicative of a fairly abnormal condition,
it's not necessarily correct to even restart execution in those cases.
For instance, `Mutex` will poison its self if the MutexGuard is dropped
while the thread is panicking. It's also usually desirable to exit the
program and let process supervision handle reinitializing the program.

This PR adds a new builder API to Arbiter to support the `stop_on_panic`
option. Future PRs can also use this builder to allow for further
customization.

Usage:
```
    let actor: Addr<SomeActor> = Arbiter::builder()
            .name("optional name")
            .stop_system_on_panic(true)
            .start(|_| SomeActor::new());
```

Or to get an `Addr<Arbiter>`:
```
    let arbiter: Addr<Arbiter> = Arbiter::builder()
            .name("optional name")
            .stop_system_on_panic(true)
            .build();
```

The existing API (`Arbiter#new`, `Arbiter#start`) uses the builder under
the hood, but does not change the behavior or API surface of the existing
API.

Resolves #110