Add Stdin and Stdout actors for working echo example
Stdin and Stdout actors are created to read from `stdin` and write to `stdout`,
respectively. Since both reading and writing from standard IO file stream and
sink, respectively, is a blocking operation, the `Stdin` and `Stdout` actors
start a separate thread for the IO and use mpsc channels as proxies for streams
and sinks. The separate threads each use a tokio runtime so the codecs and
Framed traits from teh tokio crate can be used. Thus, there are three threads:
stdin, stdout, and actix, where each thread is using a tokio runtime.

The example simply echos lines from the stdin to stdout, i.e. read from stdin
and write the line to stdout. The Stdin and Stdout actors are generically
implemented so that different codecs and/or messages can be used.