Add graceful shutdown on EOF in stdin
The tokio runtime for the Actix thread would continue to run even after the
stdin and stdout threads were gracefully stopped on an End-of-File (EOF). A
handle to the Actix "system" needed to be passed to the stdout thread to call
the `stop()` method. The stdout thread is the last thread to stop on EOF. The
stdin thread is the first stop on EOF. This full shutdown allows for all of the
logging statements to be printed before the application actually exits and
returns the user to the prompt instead of running indefinitely.