Change `Message` to `Output`
Too many things are named "message" and it gets confusing what is an actor
message and what is a "message" sent along a channel to the stdout thread. Thus,
the `Message` type is renamed to `Output`. This should hopefully make it clear
that actors can handle the output message instead of the message message.