Change to separate Input and Output messages structs
Again, hopefully to improve understanding, the Stdin actor receives "Input"
messages and the Stdout actor receives "Output" messages. Previously, the same
message (Output/Message) was used for both the Stdin and Stdout actors. This was
a little confusing in that the Stdin actor was handling "output" messages, so
a separate message type for the Stdin actor was created.