Change codec item to Bytes/BytesMut instead of String
Only the LinesCodec from the `tokio::codec` module has a `String` for the `Item`
type. The other tokio-provided codecs, BytesCodec and LengthDelimitedCodec, use
Bytes and BytesMut, respectively. In the case of the `echo` example, which just
forwards lines from stdin to stdout, Strings are okay, but to make the Stdin and
Stdout actors more generic and be able to use the other codecs without having to
change the type definitions, the items needed to be Bytes and BytesMut. Also,
Byte and BytesMut is more appropriate for stream and sink and network
communication because it is a stream of bytes, not a streams of strings.

So, a custom LinesCodec is created that wraps the tokio-provided LinesCodec and
handles conversion of String to and from Bytes/BytesMut, respectively. This
allows the encoder and decoder items for the Stdin and Stdout actors to be
`BytesMut` and `Bytes`, respectively, but still maintain the lines based
functionality of the echo example.