Remove default implementation for Stdout
If the goal is to have the Stdout actor be as generic as possible, then it
should not depend on the custom LinesCodec for the default implementation. So,
it is removed. A possible goal would be to have the Stdin and Stdout actors
included in the actix crate in the `actors` module, which includes helper
actors.