Add `Echo` actor
The `Stdin` actor is changed to be a stream handler instead of a handler of
Input messages. The Input messages are now the messages that are sent from the
Stdin actor to a recipient (subscriber). So, another actor is needed that is a
recipient of Input messages, converts the Input messages to Output messages, and
sends the Output messages to the Stdout actor. This is the `Echo` actor.

This also makes the `Stdin` actor more reusable and generic.