Add actix::io::SinkWrite<Sink>: A `Sink` IO wrapper
A `futures::sink::Sink` wrapper for writing messages into a stream
from within a Actor. It allows more the use of custom sink objects
(or the use of channels).

It's an addition to `actix::io::FramedWrite<AsyncWrite, Encoder>`
and `actix::io::Writer<AsyncWrite>`. Both of those induce copies
into the inner buffer that can be sometime avoid with custom sink.