test_actors: Make test actually run
Previous version of the test was unsound, it would spawn two futures into the
runtime and then finish successfully, not waiting for any of them to finish. We
now block the main test thread on the second message in order to ensure at
least one of them will run. We also unwrap the first result to ensure that the
message was successfully sent to the Actor

This also updates the doc example for actors::resolver accordingly

Signed-off-by: Jonathas-Conceicao <jadoliveira@inf.ufpel.edu.br>